+++ src/pages/ProjectDetailsView/index.ts
@@ -1,1 +1,2 @@
 export * from "./ProjectDetailsView";
+export * from "./ProjectAttributesChart";
